Fix format specifiers
- Fixes the following Cppcheck warnings:

  [bindings/tcl/tclrrd.c:513] (warning) %lu in format string (no. 1)
    requires 'unsigned long' but the argument type is 'signed long'.
    [invalidPrintfArgType_uint]

  [src/rrd_fetch_libdbi.c:56] (warning) %i in format string (no. 3)
    requires 'int' but the argument type is 'unsigned int'.
    [invalidPrintfArgType_sint]
  [src/rrd_fetch_libdbi.c:63] (warning) %i in format string (no. 3)
    requires 'int' but the argument type is 'unsigned int'.
    [invalidPrintfArgType_sint]
  [src/rrd_fetch_libdbi.c:85] (warning) %i in format string (no. 3)
    requires 'int' but the argument type is 'unsigned int'.
    [invalidPrintfArgType_sint]
  [src/rrd_fetch_libdbi.c:85] (warning) %i in format string (no. 4)
    requires 'int' but the argument type is 'unsigned int'.
    [invalidPrintfArgType_sint]
  [src/rrd_fetch_libdbi.c:119] (warning) %i in format string (no. 3)
    requires 'int' but the argument type is 'unsigned int'.
    [invalidPrintfArgType_sint]
  [src/rrd_fetch_libdbi.c:126] (warning) %i in format string (no. 3)
    requires 'int' but the argument type is 'unsigned int'.
    [invalidPrintfArgType_sint]
  [src/rrd_fetch_libdbi.c:148] (warning) %i in format string (no. 3)
    requires 'int' but the argument type is 'unsigned int'.
    [invalidPrintfArgType_sint]
  [src/rrd_fetch_libdbi.c:148] (warning) %i in format string (no. 4)
    requires 'int' but the argument type is 'unsigned int'.
    [invalidPrintfArgType_sint]

  [src/rrd_graph_helper.c:1504] (warning) %u in format string (no. 2)
    requires 'unsigned int' but the argument type is 'signed int'.
    [invalidPrintfArgType_uint]
  [src/rrd_graph_helper.c:1511] (warning) %u in format string (no. 2)
    requires 'unsigned int' but the argument type is 'signed int'.
    [invalidPrintfArgType_uint]
  [src/rrd_graph_helper.c:1695] (warning) %u in format string (no. 1)
    requires 'unsigned int' but the argument type is 'signed int'.
    [invalidPrintfArgType_uint]
  [src/rrd_graph_helper.c:1812] (warning) %u in format string (no. 2)
    requires 'unsigned int' but the argument type is 'signed int'.
    [invalidPrintfArgType_uint]

  [src/rrd_tune.c:392] (warning) %ld in format string (no. 3)
    requires 'long' but the argument type is 'unsigned long'.
    [invalidPrintfArgType_sint]

+++ src/rrd_graph_helper.c
@@ -1502,14 +1502,14 @@ static int parse_stack(
 
     for (i = im->gdes_c - 1; (gdp->gf == gf) && (i >= 0); i--) {
         dprintfparsed("trying to process entry %li with type %u\n", i,
-                      im->gdes[i].gf);
+                      (unsigned int) im->gdes[i].gf);
         switch (im->gdes[i].gf) {
         case GF_LINE:
         case GF_AREA:
             gdp->gf = im->gdes[i].gf;
             gdp->linewidth = im->gdes[i].linewidth;
             dprintfparsed("found matching LINE/AREA at %li with type %u\n", i,
-                          im->gdes[i].gf);
+                          (unsigned int) im->gdes[i].gf);
             break;
         default:
             break;
@@ -1692,7 +1692,7 @@ static int parse_gprint(
     }
     dprintf("VNAME : %s (%li)\n", gdp->vname, gdp->vidx);
     if ((int) gdp->cf > -1) {
-        dprintf("CF : (%u)\n", gdp->cf);
+        dprintf("CF : (%u)\n", (unsigned int) gdp->cf);
     }
     dprintf("FORMAT: \"%s\"\n", gdp->legend);
     dprintf("=================================\n");
@@ -1809,7 +1809,7 @@ static int parse_textalign(
     /* debug output */
     dprintf("=================================\n");
     dprintf("TEXTALIGN : %s\n", pa->arg_orig);
-    dprintf("ALIGNMENT : %s (%u)\n", align, gdp->txtalign);
+    dprintf("ALIGNMENT : %s (%u)\n", align, (unsigned int) gdp->txtalign);
     dprintf("=================================\n");
     /* and return */
     return 0;
